Introduce revision checks in CResultsDataManager::dumpPassFail
CResultsDataManager::dumpPassFail was assuming that there always is
a 0th revision in the results matrix and the corresponding data was
used to calculate the differences of the results vectors.
The dumpPassFail method was modified to calculate/show the differences
only if there is a 0th revision available. Also, a warning is
displayed if the 0th revision is missing.

Fixes #4