Fix an issue in the gcov based coverage reading process
Coverage data was processed by splitting the lines in gcov files
by the ':' char. In addition there is an algorithm which skips
multiline comments.

However, this algorithm was investigating only the 3rd token of the
lines i.e. it assumed that the separator character appears exactly two
times in a line. Hence, when a multiline comment were used as a single
line comment and there was at least one ':' in the comment, then the
algorithm was not able to recognize the end of the comment section
because '*/' was in the 4th, 5th, etc. token.